+++ timing.h
@@ -2,6 +2,9 @@
 #define TIMING_H
 
 #include <sys/time.h>
+#include <iostream>
+#include <chrono>
+#include <ctime>  
 
 
 /* Subtract the `struct timeval' value 'then' from 'now',
@@ -10,9 +13,9 @@
 */
 float elapsedTime(struct timeval now, struct timeval then);
 
-double currentTime();
+float currentTime();
 
 //Takes the GPU time and CPU time, compares, and prints.
-void printTimes(float gTimeCost, float cTimeCost);
+void printTimes(double gTimeCost, double cTimeCost);
 
 #endif
